# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRecordRelationDto": {
        "properties": {
          "objectCode": {
            "type": "string"
          },
          "recordId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["objectCode", "recordId"],
        "type": "object"
      },
...
      "DeleteRecordDto": {
        "properties": {
          "version": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": ["version"],
        "type": "object"
      },
      "DeleteRecordResponseDto": {
        "properties": {
          "accepted": {
            "enum": [true],
            "type": "boolean"
          }
        },
        "required": ["accepted"],
        "type": "object"
      },
...
      "PlatformAuditItemDto": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actorId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "actorName": {
            "nullable": true,
            "type": "string"
          },
          "actorType": {
            "enum": ["USER", "SYSTEM", "INTEGRATION"],
            "type": "string"
          },
          "after": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "before": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ip": {
            "nullable": true,
            "type": "string"
          },
          "reason": {
            "nullable": true,
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "resourceId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "tenantName": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "actorType",
          "action",
          "resourceType",
          "requestId",
          "createdAt"
        ],
        "type": "object"
      },
      "PlatformAuditPageDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlatformAuditItemDto"
            },
            "type": "array"
          },
          "limit": {
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": ["items", "page", "limit", "total"],
        "type": "object"
      },
...
      "RecordRelationDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "objectCode": {
            "type": "string"
          },
          "recordId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": ["id", "recordId", "objectCode", "title"],
        "type": "object"
      },
      "RecordResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ownerMemberId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "recordNo": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "type": "object"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "recordNo",
          "title",
          "values",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "RelationResultDto": {
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": ["success"],
        "type": "object"
      },
...
      "UpdateRecordDto": {
        "properties": {
          "ownerMemberId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "type": "object"
          },
          "version": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": ["version"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "crm_session": {
        "in": "cookie",
        "name": "crm_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "多租户 CRM 平台 REST API",
    "title": "Multi-tenant CRM API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{tenantCode}/audit": {
      "get": {
        "operationId": "CompanyAuditController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resourceType",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAuditPageDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["company-audit"]
      }
    },
...
    "/api/v1/workspaces/{tenantCode}/objects/{objectCode}/records/{recordId}": {
      "delete": {
        "operationId": "RecordsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRecordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRecordResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["records"]
      },
      "get": {
        "operationId": "RecordsController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["records"]
      },
      "patch": {
        "operationId": "RecordsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["records"]
      }
    },
...
    "/api/v1/workspaces/{tenantCode}/objects/{objectCode}/records/{recordId}/relations": {
      "get": {
        "operationId": "RecordRelationsController_list",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecordRelationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["record-relations"]
      },
      "post": {
        "operationId": "RecordRelationsController_add",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordRelationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["record-relations"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/objects/{objectCode}/records/{recordId}/relations/{relationId}": {
      "delete": {
        "operationId": "RecordRelationsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "relationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["record-relations"]
      }
    },
...
  },
  "servers": [],
...
}